+++ test/dummy/app/portlets/product_catalog_portlet.rb
@@ -2,7 +2,8 @@ class ProductCatalogPortlet < Cms::Portlet
 
   # Mark this as 'true' to allow the portlet's template to be editable via the CMS admin UI.
   enable_template_editor false
-     
+
+  description 'Return list of all product catalogs. [Testing portlet only. Will not be packaged with BrowserCMS]'
   def render
     # Your Code Goes Here
   end